pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "grox"
version = "0.1.0"
description = "Async-ready context and logging manager for multi-tenant services"
authors = [
  { name="Phragman", email="a@staphi.com" }
]
readme = "README.md"
license = { text = "MIT" }
dependencies = [
  "pydantic>=1.10.0,<3",
  "structlog>=23.1.0",
  "python-json-logger>=2.0.0",
  "pyyaml>=6.0",
  "click>=8.0",
  "jinja2",
  "langfabric (>=0.1.8,<0.2.0)",
  "seyaml>=0.1.1,<1.0",
  "langgraph (>=0.4.10,<0.5.0)",
  "langgraph-checkpoint-redis (>=0.0.8,<0.0.9)",
  "langchain-redis (>=0.2.3,<0.3.0)",
  "colorama (>=0.4.6,<0.5.0)",
]
requires-python = ">=3.11,<3.14"

[project.optional-dependencies]
test = [
    "pytest",
]

[project.scripts]
grox = "grox.cli:cli"

[tool.setuptools.package-data]
"grox.templates" = ["*.j2"]

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
